fix common ui test
